+++ setup.py
@@ -2,11 +2,11 @@ from setuptools import setup
 
 if __name__ == "__main__":
     setup(
-        name="pyrant",
+        name="twaddle",
         version="0.0.1",
         description="Python version of Rant v3",
         author="Chris Hengler",
         author_email="dr.c.hengler@gmail.com",
         url="https://github.com/chrishengler/PyRant",
-        packages=["pyrant"],
+        packages=["twaddle"],
     )
