+++ twaddle/tests/resources/example.dic
