+++ twaddle/tests/resources/example_with_classes.dic
